+++ qti_package_maker/common/color_wheel/__init__.py
@@ -1,6 +1,0 @@
-#!/usr/bin/env python3
-
-"""Color wheel package with legacy and next-gen implementations."""
-
-# QTI Package Maker
-from .legacy import *  # noqa: F401,F403